[Performance] first benchmarks, ScalaPropertyAccessor performance fix, more performant EvaluationContext initialization, small refactoring
