+++ engine/interpreter/src/main/scala/pl/touk/nussknacker/engine/spel/SpelExpression.scala
@@ -4,18 +4,16 @@ import java.lang.reflect.{Method, Modifier}
 import java.time.{LocalDate, LocalDateTime}
 import java.util
 import java.util.concurrent.TimeoutException
-import java.util.{Collections, Optional}
 
 import cats.data.Validated.Valid
 import cats.data.{NonEmptyList, State, StateT, Validated}
 import cats.effect.IO
 import com.typesafe.scalalogging.LazyLogging
 import org.apache.commons.lang3.StringUtils
-import org.springframework.core.convert.TypeDescriptor
 import org.springframework.expression._
 import org.springframework.expression.common.{CompositeStringExpression, LiteralExpression}
 import org.springframework.expression.spel.ast.SpelNodeImpl
-import org.springframework.expression.spel.support.{ReflectiveMethodExecutor, ReflectiveMethodResolver, StandardEvaluationContext, StandardTypeLocator}
+import org.springframework.expression.spel.support.ReflectivePropertyAccessor
 import org.springframework.expression.spel.{SpelCompilerMode, SpelEvaluationException, SpelParserConfiguration, standard}
 import pl.touk.nussknacker.engine.api
 import pl.touk.nussknacker.engine.api.Context
@@ -27,11 +25,12 @@ import pl.touk.nussknacker.engine.api.lazyy.{ContextWithLazyValuesProvider, Lazy
 import pl.touk.nussknacker.engine.api.process.ClassExtractionSettings
 import pl.touk.nussknacker.engine.api.typed.TypedMap
 import pl.touk.nussknacker.engine.api.typed.supertype.{CommonSupertypeFinder, SupertypeClassResolutionStrategy}
-import pl.touk.nussknacker.engine.api.typed.typing.{SingleTypingResult, TypedClass, TypingResult, Unknown}
+import pl.touk.nussknacker.engine.api.typed.typing.{SingleTypingResult, TypingResult, Unknown}
 import pl.touk.nussknacker.engine.dict.{KeysDictTyper, LabelsDictTyper}
 import pl.touk.nussknacker.engine.expression.NullExpression
 import pl.touk.nussknacker.engine.functionUtils.CollectionUtils
 import pl.touk.nussknacker.engine.spel.SpelExpressionParser.Flavour
+import pl.touk.nussknacker.engine.spel.internal.EvaluationContextPreparer
 
 import scala.collection.concurrent.TrieMap
 import scala.concurrent.Future
@@ -76,7 +75,7 @@ final case class ParsedSpelExpression(original: String, parser: () => Validated[
 class SpelExpression(parsed: ParsedSpelExpression,
                      expectedReturnType: TypingResult,
                      flavour: Flavour,
-                     prepareEvaluationContext: Context => EvaluationContext) extends api.expression.Expression with LazyLogging {
+                     evaluationContextPreparer: EvaluationContextPreparer) extends api.expression.Expression with LazyLogging {
 
   import pl.touk.nussknacker.engine.spel.SpelExpressionParser._
 
@@ -100,9 +99,7 @@ class SpelExpression(parsed: ParsedSpelExpression,
       return Future.successful(ValueWithLazyContext(SpelExpressionRepr(parsed.parsed, ctx, original).asInstanceOf[T], ctx.lazyContext))
     }
 
-    val evaluationContext = prepareEvaluationContext(ctx)
-    evaluationContext.setVariable(LazyValuesProviderVariableName, lazyValuesProvider)
-    evaluationContext.setVariable(LazyContextVariableName, ctx.lazyContext)
+    val evaluationContext = evaluationContextPreparer.prepareEvaluationContext(ctx, lazyValuesProvider)
 
     //TODO: async evaluation of lazy vals...
     val value = parsed.getValue[T](evaluationContext, expectedClass)
@@ -129,7 +126,7 @@ class SpelExpressionParser(parser: org.springframework.expression.spel.standard.
                            dictRegistry: DictRegistry,
                            enableSpelForceCompile: Boolean,
                            flavour: Flavour,
-                           prepareEvaluationContext: Context => EvaluationContext) extends ExpressionParser {
+                           prepareEvaluationContext: EvaluationContextPreparer) extends ExpressionParser {
 
   import pl.touk.nussknacker.engine.spel.SpelExpressionParser._
 
@@ -198,8 +195,8 @@ object SpelExpressionParser extends LazyLogging {
     parsed match {
       case e:standard.SpelExpression => forceCompile(e)
       case e:CompositeStringExpression => e.getExpressions.foreach(forceCompile)
-      case e:LiteralExpression =>
-      case e:NullExpression =>
+      case _:LiteralExpression =>
+      case _:NullExpression =>
     }
   }
 
@@ -232,236 +229,17 @@ object SpelExpressionParser extends LazyLogging {
       "distinct" -> classOf[CollectionUtils].getDeclaredMethod("distinct", classOf[util.Collection[_]]),
       "sum" -> classOf[CollectionUtils].getDeclaredMethod("sum", classOf[util.Collection[_]])
     )
-    //FIXME: configurable timeout...
-    val lazyValuesTimeout = 1 minute
     val parser = new org.springframework.expression.spel.standard.SpelExpressionParser(
       //we have to pass classloader, because default contextClassLoader can be sth different than we expect...
       new SpelParserConfiguration(SpelCompilerMode.IMMEDIATE, classLoader)
     )
-    val propertyAccessors = Seq(
-      NullPropertyAccessor, //must come first
-      new ScalaLazyPropertyAccessor(lazyValuesTimeout), // must be before scalaPropertyAccessor
-      ScalaOptionOrNullPropertyAccessor, // // must be before scalaPropertyAccessor
-      ScalaPropertyAccessor,
-      StaticPropertyAccessor,
-      MapPropertyAccessor,
-      TypedMapPropertyAccessor,
-      TypedDictInstancePropertyAccessor,
-      // it can add performance overhead so it will be better to keep it on the bottom
-      MapLikePropertyAccessor
-    )
+    val propertyAccessors = internal.propertyAccessors.configured()
 
     val classResolutionStrategy = if (strictTypeChecking) SupertypeClassResolutionStrategy.Intersection else SupertypeClassResolutionStrategy.Union
     val commonSupertypeFinder = new CommonSupertypeFinder(classResolutionStrategy, strictTypeChecking)
     val validator = new SpelExpressionValidator(new Typer(classLoader, commonSupertypeFinder, new KeysDictTyper(dictRegistry), strictMethodsChecking))
-    new SpelExpressionParser(parser, validator, dictRegistry, enableSpelForceCompile, flavour,
-      prepareEvaluationContext(classLoader, imports, propertyAccessors, functions))
-  }
-
-  private def prepareEvaluationContext[T](classLoader: ClassLoader,
-                                          expressionImports: List[String],
-                                          propertyAccessors: Seq[PropertyAccessor],
-                                          expressionFunctions: Map[String, Method])
-                                         (ctx: Context): EvaluationContext = {
-    val evaluationContext = new StandardEvaluationContext()
-    val locator = new StandardTypeLocator(classLoader)
-    expressionImports.foreach(locator.registerImport)
-    evaluationContext.setTypeLocator(locator)
-    propertyAccessors.foreach(evaluationContext.addPropertyAccessor)
-
-    evaluationContext.setMethodResolvers(optimizedMethodResolvers())
-
-    ctx.variables.foreach {
-      case (k, v) => evaluationContext.setVariable(k, v)
-    }
-    expressionFunctions.foreach {
-      case (k, v) => evaluationContext.registerFunction(k, v)
-    }
-    evaluationContext
-  }
-
-  private def optimizedMethodResolvers() : java.util.List[MethodResolver] = {
-    val mr = new ReflectiveMethodResolver {
-      override def resolve(context: EvaluationContext, targetObject: scala.Any, name: String, argumentTypes: util.List[TypeDescriptor]): MethodExecutor = {
-        val methodExecutor = super.resolve(context, targetObject, name, argumentTypes).asInstanceOf[ReflectiveMethodExecutor]
-        new OmitAnnotationsMethodExecutor(methodExecutor)
-      }
-    }
-    Collections.singletonList(mr)
-  }
-
-  object NullPropertyAccessor extends PropertyAccessor with ReadOnly {
-
-    override def getSpecificTargetClasses: Array[Class[_]] = null
-
-    override def canRead(context: EvaluationContext, target: Any, name: String): Boolean = target == null
-
-    override def read(context: EvaluationContext, target: Any, name: String): TypedValue =
-      //can we extract anything else here?
-      throw NonTransientException(name, s"Cannot invoke method/property $name on null object")
-  }
-
-  object ScalaPropertyAccessor extends PropertyAccessor with ReadOnly with Caching {
-
-    override protected def reallyFindMethod(name: String, target: Class[_]) : Option[Method] =
-      target.getMethods.find(m => m.getParameterCount == 0 && m.getName == name)
-
-
-    override protected def invokeMethod(propertyName: String, method: Method, target: Any, context: EvaluationContext): AnyRef = {
-      method.invoke(target)
-    }
-
-    override def getSpecificTargetClasses: Array[Class[_]] = null
-  }
-
-  object StaticPropertyAccessor extends PropertyAccessor with ReadOnly with StaticMethodCaching {
-
-    override protected def reallyFindMethod(name: String, target: Class[_]): Option[Method] = {
-      target.asInstanceOf[Class[_]].getMethods.find(m =>
-        m.getParameterCount == 0 && m.getName == name && Modifier.isStatic(m.getModifiers)
-      )
-    }
-
-    override protected def invokeMethod(propertyName: String, method: Method, target: Any, context: EvaluationContext): Any = {
-      method.invoke(target)
-    }
-
-    override def getSpecificTargetClasses: Array[Class[_]] = null
-  }
-
-  object ScalaOptionOrNullPropertyAccessor extends PropertyAccessor with ReadOnly with Caching {
-
-    override protected def reallyFindMethod(name: String, target: Class[_]) : Option[Method] = {
-      target.getMethods.find(m => m.getParameterCount == 0 && m.getName == name && classOf[Option[_]].isAssignableFrom(m.getReturnType))
-    }
-
-    override protected def invokeMethod(propertyName: String, method: Method, target: Any, context: EvaluationContext): Any = {
-      method.invoke(target).asInstanceOf[Option[Any]].orNull
-    }
-
-    override def getSpecificTargetClasses: Array[Class[_]] = null
-  }
-
-
-  class ScalaLazyPropertyAccessor(lazyValuesTimeout: Duration) extends PropertyAccessor with ReadOnly with Caching {
-
-    override protected def reallyFindMethod(name: String, target: Class[_]) : Option[Method] = {
-      target.getMethods.find(
-        m => m.getParameterCount == 0 &&
-        m.getReturnType == classOf[State[_,_]] &&
-        m.getName == name)
-    }
-
-    override protected def invokeMethod(propertyName: String, method: Method, target: Any, context: EvaluationContext): Any = {
-      val f = method
-        .invoke(target)
-        .asInstanceOf[StateT[IO, ContextWithLazyValuesProvider, Any]]
-      val lazyProvider = context.lookupVariable(LazyValuesProviderVariableName).asInstanceOf[LazyValuesProvider]
-      val ctx = context.lookupVariable(LazyContextVariableName).asInstanceOf[LazyContext]
-      val futureResult = f.run(ContextWithLazyValuesProvider(ctx, lazyProvider))
-      //TODO: async invocation :)
-      val (modifiedContext, value) = futureResult.unsafeRunTimed(lazyValuesTimeout)
-        .getOrElse(throw new TimeoutException(s"Timout on evaluation ${method.getDeclaringClass}:${method.getName}"))
-      context.setVariable(LazyContextVariableName, modifiedContext.context)
-      value
-    }
-
-    override def getSpecificTargetClasses: Array[Class[_]] = null
-
-  }
-
-  object MapPropertyAccessor extends PropertyAccessor with ReadOnly {
-
-    override def canRead(context: EvaluationContext, target: scala.Any, name: String): Boolean =
-      target.asInstanceOf[java.util.Map[_, _]].containsKey(name)
-
-    override def read(context: EvaluationContext, target: scala.Any, name: String) =
-      new TypedValue(target.asInstanceOf[java.util.Map[_, _]].get(name))
-
-    override def getSpecificTargetClasses = Array(classOf[java.util.Map[_, _]])
-  }
-
-  object TypedMapPropertyAccessor extends PropertyAccessor with ReadOnly {
-    //in theory this always happends, because we typed it properly ;)
-    override def canRead(context: EvaluationContext, target: scala.Any, name: String): Boolean =
-      target.asInstanceOf[TypedMap].fields.contains(name)
-
-    override def read(context: EvaluationContext, target: scala.Any, name: String) =
-      new TypedValue(target.asInstanceOf[TypedMap].fields(name))
-
-    override def getSpecificTargetClasses = Array(classOf[TypedMap])
-  }
-
-  object TypedDictInstancePropertyAccessor extends PropertyAccessor with ReadOnly {
-    //in theory this always happends, because we typed it properly ;)
-    override def canRead(context: EvaluationContext, target: scala.Any, key: String) =
-      true
-
-    // we already replaced dict's label with keys so we can just return value based on key
-    override def read(context: EvaluationContext, target: scala.Any, key: String) =
-      new TypedValue(target.asInstanceOf[DictInstance].value(key))
-
-    override def getSpecificTargetClasses = Array(classOf[DictInstance])
-  }
-
-  // mainly for avro's GenericRecord purpose
-  object MapLikePropertyAccessor extends PropertyAccessor with Caching with ReadOnly {
-
-    override protected def invokeMethod(propertyName: String, method: Method, target: Any, context: EvaluationContext): Any = {
-      method.invoke(target, propertyName)
-    }
-
-    override protected def reallyFindMethod(name: String, target: Class[_]): Option[Method] = {
-      target.getMethods.find(m => m.getName == "get" && (m.getParameterTypes sameElements Array(classOf[String])))
-    }
-
-    override def getSpecificTargetClasses: Array[Class[_]] = null
-  }
-
-  trait Caching extends CachingBase { self: PropertyAccessor =>
-
-    override def canRead(context: EvaluationContext, target: scala.Any, name: String): Boolean =
-      !target.isInstanceOf[Class[_]] && findMethod(name, target).isDefined
-
-    override protected def extractClassFromTarget(target: Any): Option[Class[_]] =
-      Option(target).map(_.getClass)
-  }
-
-  trait StaticMethodCaching extends CachingBase { self: PropertyAccessor =>
-    override def canRead(context: EvaluationContext, target: scala.Any, name: String): Boolean =
-      target.isInstanceOf[Class[_]] && findMethod(name, target).isDefined
-
-    override protected def extractClassFromTarget(target: Any): Option[Class[_]] = Option(target).map(_.asInstanceOf[Class[_]])
-  }
-
-  trait CachingBase { self: PropertyAccessor =>
-    private val methodsCache = new TrieMap[(String, Class[_]), Option[Method]]()
-
-    override def read(context: EvaluationContext, target: scala.Any, name: String): TypedValue =
-      findMethod(name, target)
-        .map { method =>
-          new TypedValue(invokeMethod(name, method, target, context))
-        }
-        .getOrElse(throw new IllegalAccessException("Property is not readable"))
-
-    protected def findMethod(name: String, target: Any): Option[Method] = {
-      //this should *not* happen as we have NullPropertyAccessor
-      val targetClass = extractClassFromTarget(target).getOrElse(throw new IllegalArgumentException(s"Null target for $name"))
-      methodsCache.getOrElseUpdate((name, targetClass), reallyFindMethod(name, targetClass))
-    }
-
-    protected def extractClassFromTarget(target: Any): Option[Class[_]]
-    protected def invokeMethod(propertyName: String, method: Method, target: Any, context: EvaluationContext): Any
-    protected def reallyFindMethod(name: String, target: Class[_]) : Option[Method]
-  }
-
-  trait ReadOnly { self: PropertyAccessor =>
-
-    override def write(context: EvaluationContext, target: scala.Any, name: String, newValue: scala.Any) =
-      throw new IllegalAccessException("Property is not writeable")
-
-    override def canWrite(context: EvaluationContext, target: scala.Any, name: String) = false
-
+    val evaluationContextPreparer = new EvaluationContextPreparer(classLoader, imports, propertyAccessors, functions)
+    new SpelExpressionParser(parser, validator, dictRegistry, enableSpelForceCompile, flavour, evaluationContextPreparer)
   }
 
 }
